+++ src/BlogBank.Service/Interfaces/IRoleService.cs
@@ -5,9 +5,9 @@ namespace BlogBank.Service.Interfaces;
 public interface IRoleService
 {
     Task<IEnumerable<Role>> GetAllAsync();
-    Task<Role?> GetByIdAsync(int id);
-    Task<bool> CodeExistsAsync(string code, int? excludeId = null);
+    Task<Role?> GetByIdAsync(long id);
+    Task<bool> CodeExistsAsync(string code, long? excludeId = null);
     Task<Role> CreateAsync(Role role);
-    Task<Role?> UpdateAsync(int id, Role role);
-    Task<bool> DeleteAsync(int id);
+    Task<Role?> UpdateAsync(long id, Role role);
+    Task<bool> DeleteAsync(long id);
 }
